+++ clients/python/flux_restful_client/client/logs.py
@@ -8,10 +8,14 @@ def main(args, parser, extra, subparser):
 
     # Update config settings on the fly
     cli.settings.update_params(args.config_params)
-    logs = cli.output(jobid=args.jobid)
-    if "Output" in logs:
-        for line in logs["Output"]:
-            # Ensure we only have one newline!
+    if args.stream:
+        for line in cli.stream_output(jobid=args.jobid, stream=args.stream):
             print(line.strip())
     else:
-        print(json.dumps(logs, indent=4))
+        logs = cli.output(jobid=args.jobid, stream=args.stream)
+        if "Output" in logs:
+            for line in logs["Output"]:
+                # Ensure we only have one newline!
+                print(line.strip())
+        else:
+            print(json.dumps(logs, indent=4))